feat: add single type image to gallery component
